% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_plus_one.R
\name{one_plus_one}
\alias{one_plus_one}
\title{Title}
\usage{
one_plus_one()
}
\value{

}
\description{
Title
}
\examples{

# This is the first demonstration
one_plus_one() ->
adding

adding + one_plus_one()


# This is the second demonstration
one_plus_one() + 2 +
4 + 8 +
7
}
