% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples_path_explode.R
\name{package_examples_exploded}
\alias{package_examples_exploded}
\title{Returns a flipbook built from package function examples}
\usage{
package_examples_exploded(
  package_name,
  package_path = ".",
  functions_path = "./R/",
  flipbook_setup = write_flipbook_setup(),
  yaml = write_xaringan_yaml(package_name),
  setup_code_chunk = write_setup_code_chunk(package_name),
  rmd = "docs/exploded_examples.Rmd",
  render = T
)
}
\arguments{
\item{setup_code_chunk}{}
}
\value{

}
\description{
Returns a flipbook built from package function examples
}
